crypto_treasury_pipeline/requirements.txt
requests
yfinance
beautifulsoup4
lxml
matplotlib
jinja2
pyyaml
pandas